Add unsignedInteger instead of bigInteger

+++ database/migrations/2023_03_30_232824_create_tasks_table.php
@@ -16,8 +16,8 @@ class CreateTasksTable extends Migration
         Schema::create('tasks', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('user_id');
-            $table->bigInteger('category_id');
-            $table->bigInteger('visitor_id')->nullable();
+            $table->unsignedInteger('category_id');
+            $table->unsignedInteger('visitor_id')->nullable();
             $table->enum('status', ['0', '1', '2'])->default('0')->comment('0 means Not Started, 1 means Completed, 2 menas In Progress');
             $table->string('name');
             $table->longText('description')->nullable();
